Audio
- SFX
- Code
- BGM

+++ Assets/Code/Runtime/Damages/IDamageable.cs
@@ -3,7 +3,7 @@ namespace SwapChains.Runtime.Entities.Damages
     public interface IDamageable
     {
         bool CanReceiveDamage();
-        void ReceiveDamage(float amount);
+        void ReceiveDamage(int amount);
         Health GetHealth();
     }
 }
